Clean-up.
git-svn-id: http://svn.tilp.info/repos/tiemu@563 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/misc/files.c
@@ -105,8 +105,9 @@ void process_buffer(gchar *buf)
 {
   gint i;
   
-  for(i=0; i<strlen(buf); i++)
-    if(buf[i]=='\r') buf[i]=' ';
+  for(i = 0; i < (int)strlen(buf); i++)
+    if(buf[i] == '\r') 
+		buf[i]=' ';
 }
 
 /*************************************/
